pyproject.toml
[project]
name = "alt"
version = "0.1.0"
description = "CLI tools for alt second brain"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0",
]

[project.scripts]
alt-db = "alt_db.cli:main"
alt-discord = "alt_discord.cli:main"
alt-body = "alt_body.cli:main"
alt-home-assistant = "alt_home_assistant.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alt_db", "src/alt_discord", "src/alt_body", "src/alt_home_assistant"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
